# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Reccomendation",
        "description": "Get All Reccomendation Books",
        "contact": {
            "email": "a.osamamahmoud96@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get all books",
                "description": "Retrieves a list of all books in the collection.",
                "operationId": "BooksList",
                "parameters": [
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter between deleted and active books [ 1 => All Active Books , 0 => All In Active Books , -1 => All Books ]",
                        "schema": {
                            "type": "integer",
                            "default": "1",
                            "enum": [
                                1,
                                0,
                                -1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of the rows per the page [ pagination ]",
                        "schema": {
                            "type": "integer",
                            "default": "10"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the desired page number [ pagination ]",
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "A Brief History of Time"
                                            },
                                            "number_of_pages": {
                                                "type": "integer",
                                                "example": 256
                                            },
                                            "is_active": {
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Add a new book",
                "description": "Adds a new book to the collection and returns the added book details.",
                "operationId": "storeBook",
                "requestBody": {
                    "description": "Pass book details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "number_of_pages"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "A Brief History of Time"
                                    },
                                    "number_of_pages": {
                                        "type": "integer",
                                        "example": 256
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book successfully added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "A Brief History of Time"
                                        },
                                        "number_of_pages": {
                                            "type": "integer",
                                            "example": 256
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You have entered this title before"
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "You have entered this interval before"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/books/restore/{id}": {
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Restore a single book",
                "description": "Restore a book from the collection given its ID.",
                "operationId": "restoreBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to restore",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful restore operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "A Brief History of Time"
                                        },
                                        "number_of_pages": {
                                            "type": "integer",
                                            "example": 256
                                        },
                                        "is_active": {
                                            "type": "boolean",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
        "/api/v1/books/recommendations/top-five": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get Top Five Recommendation books [ V1 Query Based ]",
                "description": "Retrieves a list of Top Five Recommendation books",
                "operationId": "topFiveBooks",
                "responses": {
                    "200": {
                        "description": "A List of Top Five Recommendation books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "book_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "book_name": {
                                                "type": "string",
                                                "example": "A Brief History of Time"
                                            },
                                            "num_of_read_pages": {
                                                "type": "integer",
                                                "example": 256
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            }
        },
        "/api/v1/books/recommendations/top-five/v2": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get Top Five Recommendation books [ V2 Code & loops Based ]",
                "description": "Retrieves a list of Top Five Recommendation books",
                "operationId": "topFiveBooksV2",
                "responses": {
                    "200": {
                        "description": "A List of Top Five Recommendation books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "book_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "book_name": {
                                                "type": "string",
                                                "example": "A Brief History of Time"
                                            },
                                            "num_of_read_pages": {
                                                "type": "integer",
                                                "example": 256
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            }
        },
...
    }
}